+++ packages/fdr-sdk/src/index.ts
@@ -1,5 +1,7 @@
+export * as ApiDefinition from "./api-definition";
 export * from "./client";
 export * from "./converters";
+export * as FernDocs from "./docs";
 export * as FernNavigation from "./navigation";
 export { kebabCase } from "./utils/lodash";
 export { default as once } from "./utils/lodash/once";
