--- packages/ui/docs-bundle/src/server/algolia/getSearchApiKey.ts
@@ -1,8 +1,0 @@
-import { algoliaApiKey, provideAlgoliaClient } from "./service";
-
-export function getSearchApiKey(indexSegmentId: string): string | undefined {
-    return provideAlgoliaClient().generateSecuredApiKey(algoliaApiKey(), {
-        filters: `indexSegmentId:${indexSegmentId}`,
-        validUntil: Math.floor(Date.now() / 1_000) + 60 * 60 * 24,
-    });
-}
