+++ packages/ui/fern-docs-search-server/.stylelintrc.json
@@ -1,0 +1,1 @@
+{ "extends": ["../../../shared/stylelintrc.shared.json"] }
