+++ packages/ui/fern-docs-search-server/src/algolia/getSearchApiKey.ts
@@ -1,0 +1,12 @@
+import { algoliaApiKey, algoliaSearchIndex, provideAlgoliaClient } from "./service";
+
+export async function getSearchApiKey(indexSegmentId: string): Promise<string | undefined> {
+    return provideAlgoliaClient().generateSecuredApiKey({
+        parentApiKey: algoliaApiKey(),
+        restrictions: {
+            filters: `indexSegmentId:${indexSegmentId}`,
+            validUntil: Math.floor(Date.now() / 1_000) + 60 * 60 * 24,
+            restrictIndices: [algoliaSearchIndex()],
+        },
+    });
+}
