+++ packages/ui/fern-docs-search-server/src/algolia/index.ts
@@ -1,0 +1,2 @@
+export * from "./getSearchApiKey";
+export * from "./records";
