+++ packages/ui/fern-docs-search-server/src/algolia/records/index.ts
@@ -1,0 +1,1 @@
+export * from "./generateAlgoliaRecords";
