+++ packages/ui/fern-docs-search-server/src/algolia/records/types.ts
@@ -1,0 +1,3 @@
+export interface GenerateAlgoliaRecordsFlags {
+    isFieldRecordsEnabled: boolean;
+}
